delete import and rename files
